reversion code of PRSC into the prsc main file
